refactor: update import for hooks; update calls for useCurrentUser; login after refresh
