+++ src/components/header/header.jsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 import { FaBars, FaTimes } from "react-icons/fa";
 import { Link } from "react-router-dom";
 import { GET_USER_AUTHENTICATION } from "../../graphql/get-user-authentication";
-import { useModals } from "../../hooks";
+import { useClient, useCurrentUser, useModals } from "../../hooks";
 import headerArrow from "../../images/global-header-arrow.svg";
 import headerBook from "../../images/global-header-book.svg";
 import headerDoor from "../../images/global-header-door.svg";
@@ -15,17 +15,12 @@ import headerStar from "../../images/global-header-star.svg";
 import { Button, LoginModal, SignupModal } from "../index";
 
 export const Header = () => {
-  const [token, setToken] = useState("");
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const { currentUser, setCurrentUser } = useContext(CurrentUserContext);
+  const { setToken } = useClient();
+  const { currentUser, setCurrentUser, isLoggedIn, setIsLoggedIn } =
+    useCurrentUser();
   const { data, loading, error } = useQuery(GET_USER_AUTHENTICATION);
   const { registerModal, openModal, closeModal } = useModals();
   const [open, setOpen] = useState(false);
-  const jwt = localStorage.getItem("jwt");
-
-  // const isLoggedIn = useMemo(() => {
-  //   return true ? token !== "" : false;
-  // }, [token]);
 
   const handleMenu = () => {
     setOpen((prev) => !prev);
@@ -37,21 +32,17 @@ export const Header = () => {
   }, []);
 
   useEffect(() => {
-    if (jwt) {
-      setToken(jwt);
-    }
-
-    if (data && data.authenticatedItem) {
+    if (!loading && data && data.authenticatedItem) {
       setCurrentUser(data.authenticatedItem);
       setIsLoggedIn(true);
     }
-    //needed to change value of isLoggedIn when a user logs in
-    //logging in happens in login-modal.jsx and
 
-    if (Object.values(currentUser).length > 0) {
-      setIsLoggedIn(true);
+    if (!loading && data && data.authenticatedItem === null) {
+      localStorage.removeItem("jwt");
+      setToken(null);
     }
-  }, [jwt, data, currentUser]);
+  }, [data]);
+
   return (
     <header className="flex flex-row w-full relative z-10 pl-5 pr-5 mx-auto justify-between items-center text-lg font-arialRegular x-0 content-center h-20 bg-black text-stone-50 xs:pl-2 xs:pr-2 md:text-md md:pr-2 md:pl-2">
       <Header.Logo />
@@ -223,7 +214,7 @@ Header.NavLinks = ({
 };
 
 Header.Profile = () => {
-  const { setToken } = useClient();
+  const { setToken, token } = useClient();
   const { currentUser, setCurrentUser, isLoggedIn, setIsLoggedIn } =
     useCurrentUser();
   const { data, loading, error } = useQuery(GET_USER_AUTHENTICATION);
