+++ src/store/client-context-provider.jsx
@@ -4,38 +4,66 @@ import { useEffect, useState } from "react";
 import { ClientContext } from "./client-context";
 
 export const ClientContextProvider = ({ children }) => {
-  const [token, setToken] = useState(null);
+  const [token, setToken] = useState(localStorage.getItem("jwt"));
+  const [client, setClient] = useState(null);
+
+  // useEffect(() => {
+  //   setToken(localStorage.getItem("jwt"));
+  // }, []);
 
   useEffect(() => {
-    setToken(localStorage.getItem("jwt"));
+    const httpLink = createHttpLink({
+      uri:
+        //in Vite, use special object `import.meta.env` to access enviroment variables
+        import.meta.env.MODE === "production"
+          ? //created environment variables must be prefixed by VITE
+            import.meta.env.VITE_API_URI
+          : "https://slopopedia-api-a5fe9aef64e8.herokuapp.com/api/graphql",
+    });
+
+    const authLink = setContext((_, { headers }) => {
+      return {
+        headers: {
+          ...headers,
+          authorization: token ? `Bearer ${token}` : "",
+        },
+      };
+    });
+    setClient(
+      new ApolloClient({
+        link: authLink.concat(httpLink),
+        cache: new InMemoryCache(),
+      })
+    );
   }, []);
 
-  const httpLink = createHttpLink({
-    uri:
-      //in Vite, use special object `import.meta.env` to access enviroment variables
-      import.meta.env.MODE === "production"
-        ? //created environment variables must be prefixed by VITE
-          import.meta.env.VITE_API_URI
-        : "https://slopopedia-api-a5fe9aef64e8.herokuapp.com/api/graphql",
-  });
+  useEffect(() => {
+    if (client) {
+      const httpLink = createHttpLink({
+        uri:
+          //in Vite, use special object `import.meta.env` to access enviroment variables
+          import.meta.env.MODE === "production"
+            ? //created environment variables must be prefixed by VITE
+              import.meta.env.VITE_API_URI
+            : "https://slopopedia-api-a5fe9aef64e8.herokuapp.com/api/graphql",
+      });
 
-  const authLink = setContext((_, { headers }) => {
-    return {
-      headers: {
-        ...headers,
-        authorization: token ? `Bearer ${token}` : "",
-      },
-    };
-  });
+      const authLink = setContext((_, { headers }) => {
+        return {
+          headers: {
+            ...headers,
+            authorization: token ? `Bearer ${token}` : "",
+          },
+        };
+      });
+      console.log(client);
+      client.setLink(authLink.concat(httpLink));
+    }
+  }, [token]);
 
-  const client = new ApolloClient({
-    link: authLink.concat(httpLink),
-    cache: new InMemoryCache(),
-  });
-
-  return (
+  return client ? (
     <ClientContext.Provider value={{ token, setToken, client }}>
       {children}
     </ClientContext.Provider>
-  );
+  ) : null;
 };
